bholu-ai/requirements.txt
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
requests==2.31.0
jsonschema==4.22.0
rich==13.7.1
hypothesis==6.100.1
pytest==8.2.0
pytest-mock==3.14.0
